Move storage types to allocate/storage/types
