+++ packages/codec/lib/evm/types.ts
@@ -2,6 +2,7 @@ import * as Common from "@truffle/codec/common";
 import * as Storage from "@truffle/codec/storage/types";
 import * as Ast from "@truffle/codec/ast";
 import * as Allocation from "@truffle/codec/allocate/types";
+import { StorageAllocations } from "@truffle/codec/allocate/storage/types";
 import {
   AbiAllocations,
   CalldataAllocations,
@@ -37,7 +38,7 @@ export interface EvmInfo {
 }
 
 export interface AllocationInfo {
-  storage?: Allocation.StorageAllocations;
+  storage?: StorageAllocations;
   memory?: Allocation.MemoryAllocations;
   abi?: AbiAllocations;
   calldata?: CalldataAllocations;
